+++ examples/gpu-particles/FragmentShaderParticles.cpp
@@ -98,8 +98,8 @@ void FragmentShaderParticles::initialize()
 void FragmentShaderParticles::reset()
 {
     // Choose appropriate width and height for the current number of particles
-    int size = m_positions.size();
-    m_width  = (int)sqrt((float)size);
+    int size = static_cast<int>(m_positions.size());
+    m_width  = static_cast<int>(sqrt(static_cast<float>(size)));
     m_height = m_width;
     int remain = size - (m_height * m_width);
     m_height += remain / m_width + (remain % m_width == 0 ? 0 : 1);
